+++ internal/tools/trimspec/main.go
@@ -38,6 +38,8 @@ var wantedSchemas = []string{
 	"CSVImportError",
 	"KnowledgeLimitsResponse",
 	"FieldDefinition",
+	"ScheduledJobResponse",
+	"TaskInvokeResponse",
 }
 
 func main() {